Switch to a two-bit event to enable fast set
This let us skip calling into plc and contending for the plc queue lock when
there is no contention on the event, at the cost of greatly complicating the
code.

Instead of using a single bit to indicate whether or not the event was available
(the previous `AtomicBool`) we now use two bits to indicate whether or not the
the event is available and whether or not there is another thread parked or
trying to park while waiting for the event to become available.

When this latter bit is unset, we can bypass plc altogether and just use atomic
writes to set or obtain the event. When it is set, we need to synchronize with
the waiting threads to make sure we don't end up setting the event while
ignoring a waiting thread or obtaining the event without making sure the
AVAILABLE bit remains consistent.