Optimize some atomic memory operations
Replace `AtomicU8::fetch_xxx()` with `AtomicU8::swap()` or `AtomicU8::store()`
where possible. These are faster because they're loop-free and branch-free.

Changing this for the timeout handler in `suspend_one()` requires checking if
the event became available after we were unparked (because of timeout) and
bubbling that result back to the caller.