Update examples/readme.rs to feel more complete
This revised approach demonstrates the usage of `Awaitable::wait_for()` with
timeouts, and demonstrates a reasonable usage that isn't wanting for
`WaitForMultipleObject` support with worker thread aborts handled via a timeout.

The consistency and atomicity guarantees of `AutoResetEvent` are allowed to
shine by using raw memory access rather than wrapping `ThreadMessage` in an
`RwLock`.